Remove non-functional PUT method
Re discussion on [listmonk issue #1012](https://github.com/knadh/listmonk/issues/1012):

The endpoint `/api/subscribers/lists/:id` doesn't seem to be functionally different from the endpoint for [modifying multiple users](https://listmonk.app/docs/apis/subscribers/#put-apisubscriberslists). It returns `{"message":"No IDs given."}`, even though the `id` is given in the url. It works if `"ids":[1]` (for example) is included in the payload, which in that case simply makes it identical to the method for multiple users.